requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
flake8==3.9.2
idna==3.2
mccabe==0.6.1
Pillow==8.3.1
pycodestyle==2.7.0
pyflakes==2.3.1
pytesseract==0.3.8
requests==2.26.0
selenium==3.141.0
urllib3==1.26.6
